Use Kubernetes own API to process KUBECONFIG env-var
This removes the explicit handling of KUBECONFIG, in favour of letting
the vendored k8s library manage it.

This brings in support for KUBECONFIG containing a list of configuration
files to be merged, and removes hardcoded default values.

I also ran `dep ensure` as this removes the only user of
`github.com/docker/docker/pkg/homedir`, which was pulling in
`github.com/docker/docker/pkg/idtools` and transitively a bunch of
otherwise-unneeded stuff.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>